Generalize contracts and services for sharing
