+++ app/contracts/shares/delete_contract.rb
@@ -1,8 +1,6 @@
-# frozen_string_literal: true
-
 # -- copyright
 # OpenProject is an open source project management software.
-# Copyright (C) 2023 the OpenProject GmbH
+# Copyright (C) 2010-2023 the OpenProject GmbH
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License version 3.
@@ -28,17 +26,14 @@
 # See COPYRIGHT and LICENSE files for more details.
 # ++
 
-module WorkPackageMembers::Concerns::RoleAssignment
-  include Members::Concerns::RoleAssignment
+module Shares
+  class DeleteContract < ::DeleteContract
+    validate :member_is_deletable
 
-  # Work package memberships have a unique distinction from
-  # project memberships. A User should be able to be granted
-  # "Role X" independently and via a group. Meaning that for role assignment
-  # as compared to Project memberships, the existing roles we want to take
-  # into account are those that have not been inherited.
-  def existing_ids
-    model.member_roles
-         .select { _1.inherited_from.nil? }
-         .map(&:role_id)
+    private
+
+    def member_is_deletable
+      errors.add(:base, :not_deletable) unless model.some_roles_deletable?
+    end
   end
 end
