+++ app/services/shares/set_attributes_service.rb
@@ -26,28 +26,20 @@
 # See COPYRIGHT and LICENSE files for more details.
 # ++
 
-class WorkPackageMembers::UpdateService < BaseServices::Update
-  include Members::Concerns::CleanedUp
+module Shares
+  class SetAttributesService < ::BaseServices::SetAttributes
+    prepend Shares::Concerns::RoleAssignment
 
-  protected
+    private
 
-  def after_perform(service_call)
-    return service_call unless service_call.success?
+    def set_attributes(params)
+      super
 
-    work_package_member = service_call.result
-
-    update_group_roles(work_package_member) if work_package_member.principal.is_a?(Group)
-
-    service_call
-  end
-
-  def update_group_roles(work_package_member)
-    Groups::UpdateRolesService
-      .new(work_package_member.principal,
-           current_user: user,
-           contract_class: EmptyContract)
-      .call(member: work_package_member,
-            send_notifications: false,
-            message: nil)
+      model.change_by_system do
+        if model.entity.respond_to?(:project)
+          model.project = model.entity&.project
+        end
+      end
+    end
   end
 end
