+++ spec/contracts/shares/work_packages/create_contract_spec.rb
@@ -29,7 +29,7 @@
 require "spec_helper"
 require_relative "shared_contract_examples"
 
-RSpec.describe WorkPackageMembers::CreateContract do
+RSpec.describe Shares::WorkPackages::CreateContract do
   it_behaves_like "work package member contract" do
     let(:member) do
       Member.new(roles: member_roles,
