+++ spec/contracts/shares/work_packages/shared_contract_examples.rb
