+++ spec/contracts/shares/work_packages/update_contract_spec.rb
@@ -29,7 +29,7 @@
 require "spec_helper"
 require_relative "shared_contract_examples"
 
-RSpec.describe WorkPackageMembers::UpdateContract do
+RSpec.describe Shares::WorkPackages::UpdateContract do
   it_behaves_like "work package member contract" do
     let(:member) do
       build_stubbed(:work_package_member,
